Remove duplicated methods and modularize method
